requirements.txt
supabase==1.0.3
python-dotenv==1.0.0
requests==2.31.0
google-auth-oauthlib==1.0.0
google-api-python-client==2.86.0
